Move test to TestWebSocket
